extract common part
